Fix direct Sean review items
