+++ tests/unit/test_textUtils.py
@@ -6,15 +6,9 @@
 """Unit tests for the textUtils module."""
 
 import unittest
-from collections.abc import Callable
-from types import SimpleNamespace
-from typing import Any
-from unittest.mock import Mock, patch
 
-from textUtils import UnicodeNormalizationOffsetConverter, WideStringOffsetConverter, WordSegmenter
+from textUtils import UnicodeNormalizationOffsetConverter, WideStringOffsetConverter
 from textUtils.uniscribe import splitAtCharacterBoundaries
-from textUtils.segFlag import WordSegFlag
-from textUtils.wordSeg.wordSegUtils import WordSegWithSeparatorOffsetConverter
 
 FACE_PALM = "\U0001f926"  # 🤦
 SMILE = "\U0001f60a"  # 😊
@@ -448,253 +442,3 @@ class TestUniscribeSplitAtCharacterBoundaries(unittest.TestCase):
 
 	def test_hebrew(self):
 		self._testHelper("בְּרֵאשִׁית", ["בְּ", "רֵ", "א", "שִׁ", "י", "ת"])
-
-
-class TestChineseWordSegmentationInitialization(unittest.TestCase):
-	def _makeMockJiebaDll(self) -> SimpleNamespace:
-		return SimpleNamespace(
-			initJieba=Mock(return_value=True),
-			calculateWordOffsets=Mock(),
-			insertUserWord=Mock(),
-			deleteUserWord=Mock(),
-			find=Mock(),
-			freeOffsets=Mock(),
-		)
-
-	def _setWordSegConfig(self, *, initForUnusedLang: bool) -> Callable[[], None]:
-		import config
-		from config.featureFlag import FeatureFlag
-		from config.featureFlagEnums import WordNavigationUnitFlag
-
-		originalInitForUnusedLang = config.conf["documentNavigation"]["initWordSegForUnusedLang"]
-		originalWordSegmentationStandard = config.conf["documentNavigation"]["wordSegmentationStandard"]
-		config.conf["documentNavigation"]["initWordSegForUnusedLang"] = initForUnusedLang
-		config.conf["documentNavigation"]["wordSegmentationStandard"] = FeatureFlag(
-			WordNavigationUnitFlag.AUTO,
-			behaviorOfDefault=WordNavigationUnitFlag.AUTO,
-		)
-
-		def restoreConfig() -> None:
-			config.conf["documentNavigation"]["initWordSegForUnusedLang"] = originalInitForUnusedLang
-			config.conf["documentNavigation"]["wordSegmentationStandard"] = originalWordSegmentationStandard
-
-		return restoreConfig
-
-	def test_doesNotInitializeForUnusedLanguageByDefault(self) -> None:
-		from textUtils.wordSeg.wordSegStrategy import ChineseWordSegmentationStrategy
-
-		originalLib = ChineseWordSegmentationStrategy._lib
-		restoreConfig = self._setWordSegConfig(initForUnusedLang=False)
-		ChineseWordSegmentationStrategy._lib = None
-		try:
-			with (
-				patch.object(ChineseWordSegmentationStrategy, "isUsingRelatedLanguage", return_value=False),
-				patch("textUtils.wordSeg.wordSegStrategy.cdll.LoadLibrary") as loadLibrary,
-			):
-				ChineseWordSegmentationStrategy._initCppJieba()
-
-			loadLibrary.assert_not_called()
-		finally:
-			ChineseWordSegmentationStrategy._lib = originalLib
-			restoreConfig()
-
-	def test_initializesForUnusedLanguageWhenConfigured(self) -> None:
-		from textUtils.wordSeg.wordSegStrategy import ChineseWordSegmentationStrategy
-
-		mockDll = self._makeMockJiebaDll()
-		originalLib = ChineseWordSegmentationStrategy._lib
-		restoreConfig = self._setWordSegConfig(initForUnusedLang=True)
-		ChineseWordSegmentationStrategy._lib = None
-		try:
-			with (
-				patch.object(ChineseWordSegmentationStrategy, "isUsingRelatedLanguage", return_value=False),
-				patch(
-					"textUtils.wordSeg.wordSegStrategy.cdll.LoadLibrary",
-					return_value=mockDll,
-				) as loadLibrary,
-			):
-				ChineseWordSegmentationStrategy._initCppJieba()
-
-			loadLibrary.assert_called_once()
-			mockDll.initJieba.assert_called_once()
-		finally:
-			ChineseWordSegmentationStrategy._lib = originalLib
-			restoreConfig()
-
-	def test_forceInitStillInitializesForUnusedLanguage(self) -> None:
-		from textUtils.wordSeg.wordSegStrategy import ChineseWordSegmentationStrategy
-
-		mockDll = self._makeMockJiebaDll()
-		originalLib = ChineseWordSegmentationStrategy._lib
-		restoreConfig = self._setWordSegConfig(initForUnusedLang=False)
-		ChineseWordSegmentationStrategy._lib = None
-		try:
-			with (
-				patch.object(ChineseWordSegmentationStrategy, "isUsingRelatedLanguage", return_value=False),
-				patch(
-					"textUtils.wordSeg.wordSegStrategy.cdll.LoadLibrary",
-					return_value=mockDll,
-				) as loadLibrary,
-			):
-				ChineseWordSegmentationStrategy._initCppJieba(forceInit=True)
-
-			loadLibrary.assert_called_once()
-			mockDll.initJieba.assert_called_once()
-		finally:
-			ChineseWordSegmentationStrategy._lib = originalLib
-			restoreConfig()
-
-	def test_initFailureDisablesCppJieba(self) -> None:
-		from textUtils.wordSeg.wordSegStrategy import ChineseWordSegmentationStrategy
-
-		mockDll = self._makeMockJiebaDll()
-		mockDll.initJieba.return_value = False
-		originalLib = ChineseWordSegmentationStrategy._lib
-		restoreConfig = self._setWordSegConfig(initForUnusedLang=False)
-		ChineseWordSegmentationStrategy._lib = None
-		try:
-			with (
-				patch.object(ChineseWordSegmentationStrategy, "isUsingRelatedLanguage", return_value=False),
-				patch(
-					"textUtils.wordSeg.wordSegStrategy.cdll.LoadLibrary",
-					return_value=mockDll,
-				) as loadLibrary,
-				patch("textUtils.wordSeg.wordSegStrategy.log.debugWarning") as debugWarning,
-			):
-				ChineseWordSegmentationStrategy._initCppJieba(forceInit=True)
-
-			loadLibrary.assert_called_once()
-			mockDll.initJieba.assert_called_once()
-			self.assertIsNone(ChineseWordSegmentationStrategy._lib)
-			debugWarning.assert_called_once()
-			self.assertIn("Failed to initialize cppjieba", debugWarning.call_args.args[0])
-		finally:
-			ChineseWordSegmentationStrategy._lib = originalLib
-			restoreConfig()
-
-
-class TestWordSegmenter(unittest.TestCase):
-	"""Tests for the WordSegmenter class."""
-
-	def test_basicLatin(self) -> None:
-		text = "hello world"
-		segmenter = WordSegmenter(text, wordSegFlag=WordSegFlag.UNISCRIBE)
-		self.assertEqual(segmenter.getSegmentForOffset(0), (0, 6))
-		self.assertEqual(segmenter.getSegmentForOffset(5), (0, 6))
-		self.assertEqual(segmenter.getSegmentForOffset(6), (6, 11))
-		self.assertEqual(segmenter.getSegmentForOffset(11), (6, 11))
-
-	def test_chinese(self) -> None:
-		text = "你好世界"
-
-		from textUtils.wordSeg.wordSegStrategy import ChineseWordSegmentationStrategy
-
-		ChineseWordSegmentationStrategy._initCppJieba(forceInit=True)
-		segmenter = WordSegmenter(text, wordSegFlag=WordSegFlag.CHINESE)
-		self.assertEqual(segmenter.getSegmentForOffset(0), (0, 2))
-		self.assertEqual(segmenter.getSegmentForOffset(1), (0, 2))
-		self.assertEqual(segmenter.getSegmentForOffset(2), (2, 4))
-		self.assertEqual(segmenter.getSegmentForOffset(3), (2, 4))
-		self.assertEqual(segmenter.getSegmentForOffset(4), (2, 4))
-
-	def test_chineseSegmentationFailureStoresEmptyWordEnds(self) -> None:
-		from textUtils.wordSeg.wordSegStrategy import ChineseWordSegmentationStrategy
-
-		mockDll = SimpleNamespace(
-			calculateWordOffsets=Mock(return_value=False),
-			freeOffsets=Mock(),
-		)
-		originalLib = ChineseWordSegmentationStrategy._lib
-		ChineseWordSegmentationStrategy._lib = mockDll
-		try:
-			strategy = ChineseWordSegmentationStrategy("你好世界")
-			self.assertEqual(strategy.wordEnds, [])
-			self.assertEqual(strategy.segmentedText(), "你好世界")
-		finally:
-			ChineseWordSegmentationStrategy._lib = originalLib
-
-
-class TestWordSegInitialize(unittest.TestCase):
-	def test_runsAllRegisteredInitializers(self) -> None:
-		from textUtils import wordSeg
-		from textUtils.wordSeg import wordSegStrategy
-
-		calls: list[str] = []
-
-		def firstInitializer() -> None:
-			calls.append("first")
-
-		def secondInitializer() -> None:
-			calls.append("second")
-
-		class ImmediateThread:
-			def __init__(
-				self,
-				target: Callable[..., Any],
-				args: tuple[Any, ...] | None = None,
-				kwargs: dict[str, Any] | None = None,
-				daemon: bool = False,
-				name: str | None = None,
-			) -> None:
-				self.target: Callable[..., Any] = target
-				self.args: tuple[Any, ...] = () if args is None else args
-				self.kwargs: dict[str, Any] = {} if kwargs is None else kwargs
-				self.daemon: bool = daemon
-				self.name: str | None = name
-
-			def start(self) -> None:
-				self.target(*self.args, **self.kwargs)
-
-		initializerList: list[tuple[str, str, Callable[..., Any], tuple[Any, ...], dict[str, Any]]] = [
-			("missingModule", "firstInitializer", firstInitializer, (), {}),
-			("missingModule", "secondInitializer", secondInitializer, (), {}),
-		]
-		with (
-			patch.object(wordSegStrategy, "_initializerList", initializerList),
-			patch("threading.Thread", ImmediateThread),
-		):
-			wordSeg.initialize()
-
-		self.assertEqual(calls, ["first", "second"])
-
-
-class TestWordSegWithSeparatorOffsetConverter(unittest.TestCase):
-	def _makeConverter(self) -> WordSegWithSeparatorOffsetConverter:
-		def segmentedText(sep: str, newSepIndex: list[int]) -> str:
-			newSepIndex.append(2)
-			return "ab cd"
-
-		with patch(
-			"textUtils.wordSeg.wordSegUtils.WordSegmenter",
-			return_value=SimpleNamespace(segmentedText=segmentedText),
-		):
-			return WordSegWithSeparatorOffsetConverter("abcd")
-
-	def test_strToEncodedOffsetsMapsEndAndClampsOutOfRange(self) -> None:
-		converter = self._makeConverter()
-
-		self.assertEqual(converter.strToEncodedOffsets(-1), 0)
-		self.assertEqual(converter.strToEncodedOffsets(2, 4), (3, 5))
-		self.assertEqual(converter.strToEncodedOffsets(4), 5)
-		self.assertEqual(converter.strToEncodedOffsets(5, 6), (5, 5))
-
-	def test_strToEncodedOffsetsRaisesOnOutOfRangeWhenRequested(self) -> None:
-		converter = self._makeConverter()
-
-		with self.assertRaises(IndexError):
-			converter.strToEncodedOffsets(5, raiseOnError=True)
-
-	def test_encodedToStrOffsetsMapsEndAndClampsOutOfRange(self) -> None:
-		converter = self._makeConverter()
-
-		self.assertEqual(converter.encodedToStrOffsets(-1), 0)
-		self.assertEqual(converter.encodedToStrOffsets(2), 2)
-		self.assertEqual(converter.encodedToStrOffsets(2, 3), (2, 2))
-		self.assertEqual(converter.encodedToStrOffsets(5), 4)
-		self.assertEqual(converter.encodedToStrOffsets(6, 7), (4, 4))
-
-	def test_encodedToStrOffsetsRaisesOnOutOfRangeWhenRequested(self) -> None:
-		converter = self._makeConverter()
-
-		with self.assertRaises(IndexError):
-			converter.encodedToStrOffsets(6, raiseOnError=True)
