+++ user_docs/en/changes.md
@@ -7,7 +7,7 @@
 ### New Features
 
 * Chinese text can now be navigated by word using built-in input gestures.
-Several GUI elements were added to configure this in the "Document Navigation" panel. (#18735, @CrazySteve0605)
+  A Word Segmentation Standard setting was added to the "Document Navigation" panel. (#18735, @CrazySteve0605)
 * Braille output for Chinese now includes spaces between words. (#18865, @CrazySteve0605)
 * The braille "word wrap" option has been replaced with a three-valued "Text wrap" option: Off, Show mark when words are cut, and At word boundaries. (#17010, @LeonarddeR)
   * In modes that show a continuation mark, when a word is cut across rows, the last cell of the row now shows a continuation mark (braille dots 7-8) so it is clear that the word continues on the next row.
@@ -27,6 +27,8 @@ Please refer to [the developer guide](https://download.nvaccess.org/documentatio
 #### Deprecations
 
 * The `braille.wordWrap` configuration key is deprecated and bridged to `braille.textWrap`. (#17010, @LeonarddeR)
+* The `useUniscribe` attribute of `textInfos.offsets.OffsetsTextInfo` and its subclasses is deprecated.
+  Use `charSegFlag` and `wordSegFlag` instead. (#18735)
 
 <!-- Beyond this point, Markdown should not be automatically linted, as we don't modify old change log sections and lint rules may change over time. -->
 <!-- markdownlint-disable -->
@@ -147,9 +149,6 @@ Use the individual test commands instead: `runcheckpot.bat`, `rununittests.bat`,
 * The `speechDictHandler.ENTRY_TYPE_*` constants are deprecated.
 Use the `speechDictHandler.types.EntryType` enumeration instead. (#19430, @LeonarddeR)
 * `speechDictHandler.SpeechDictEntry` and `speechDictHandler.SpeechDict` have been moved to `speechDictHandler.types`. (#19430, @LeonarddeR)
-* `useUniscribe` from `textUtils.offset.OffsetsTextInfo` and its subclasses is deprecated.
-  Use `charSegFlag` and `wordSegFlag` instead. (#18735)
-
 ## 2026.1.1
 
 This is a patch release to fix security issues.
